+++ tests/test_error_tools.py
@@ -3,13 +3,23 @@ from __future__ import unicode_literals
 import errno
 import unittest
 
+import fs.errors
 from fs.error_tools import convert_os_errors
-from fs import errors as fserrors
 
 
 class TestErrorTools(unittest.TestCase):
-    def assert_convert_os_errors(self):
+    def test_convert_enoent(self):
+        exception = OSError(errno.ENOENT, "resource not found")
+        with self.assertRaises(fs.errors.ResourceNotFound) as ctx:
+            with convert_os_errors("stat", "/tmp/test"):
+                raise exception
+        self.assertEqual(ctx.exception.exc, exception)
+        self.assertEqual(ctx.exception.path, "/tmp/test")
 
-        with self.assertRaises(fserrors.ResourceNotFound):
-            with convert_os_errors("foo", "test"):
-                raise OSError(errno.ENOENT)
+    def test_convert_enametoolong(self):
+        exception = OSError(errno.ENAMETOOLONG, "File name too long: test")
+        with self.assertRaises(fs.errors.PathError) as ctx:
+            with convert_os_errors("stat", "/tmp/test"):
+                raise exception
+        self.assertEqual(ctx.exception.exc, exception)
+        self.assertEqual(ctx.exception.path, "/tmp/test")